server/requirements.txt
Flask==0.10.1
Werkzeug==0.11.11
python-dotenv==0.6.0
flask_mail==0.9.1
flask_mongoengine==0.8
validators==0.11.0
flask-s3==0.3.1